add rbac

+++ app/resources/topic.py
@@ -1,6 +1,8 @@
 from flask_restful import Resource
 from flask import request
+from flask_jwt_extended import jwt_required, get_jwt_identity
 from app import db
+from .auth import check_if_user_is_admin
 from app.models import Topic
 
 
@@ -8,14 +10,15 @@ class TopicsGETResource(Resource):
     def get(self):
         return [topic.serialize() for topic in Topic.query.all()]
     
-class TopicResource(Resource):
-    def get(self, id):
-        topic = Topic.query.get(id)
-        if topic:
-            return topic.serialize()
-        return None
-    
+class TopicsPOSTResource(Resource):
     def post(self):
+        result = jwt_required()(self._post)()  # Applying decorator directly and calling wrapped method
+        return result
+
+    def _post(self):
+        user_id = get_jwt_identity()
+        if not check_if_user_is_admin(user_id):
+            return {"message": "You are not authorized to perform this action"}, 403
         data = request.get_json()
         name = data.get('name')
         description = data.get('description')
@@ -24,7 +27,18 @@ class TopicResource(Resource):
         db.session.commit()
         return new_topic.serialize(), 201
     
+class TopicResource(Resource):
+    def get(self, id):
+        topic = Topic.query.get(id)
+        if topic:
+            return topic.serialize()
+        return None
+    
+    @jwt_required
     def put(self, id):
+        user_id = get_jwt_identity()
+        if not check_if_user_is_admin(user_id):
+            return {"message": "You are not authorized to perform this action"}, 403
         data = request.get_json()
         name = data.get('name')
         description = data.get('description')
@@ -35,8 +49,11 @@ class TopicResource(Resource):
             db.session.commit()
             return topic.serialize()
         return None
-    
+    @jwt_required
     def delete(self, id):
+        user_id = get_jwt_identity()
+        if not check_if_user_is_admin(user_id):
+            return {"message": "You are not authorized to perform this action"}, 403
         topic = Topic.query.get(id)
         if topic:
             db.session.delete(topic)
